+++ plugins/stock-photo/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @brains/stock-photo
 
+## 0.2.0-alpha.409
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/image@0.2.0-alpha.409
+  - @brains/utils@0.2.0-alpha.409
+  - @brains/plugins@0.2.0-alpha.409
+
+## 0.2.0-alpha.408
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/image@0.2.0-alpha.408
+  - @brains/utils@0.2.0-alpha.408
+  - @brains/plugins@0.2.0-alpha.408
+
 ## 0.2.0-alpha.407
 
 ### Patch Changes