+++ shared/console-theme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @brains/console-theme
 
+## 0.2.0-alpha.409
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/theme-base@0.2.0-alpha.409
+
+## 0.2.0-alpha.408
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/theme-base@0.2.0-alpha.408
+
 ## 0.2.0-alpha.407
 
 ### Patch Changes