+++ shared/content-formatters/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @brains/content-formatters
 
+## 0.2.0-alpha.409
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.409
+  - @brains/utils@0.2.0-alpha.409
+
+## 0.2.0-alpha.408
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.408
+  - @brains/utils@0.2.0-alpha.408
+
 ## 0.2.0-alpha.407
 
 ### Patch Changes