+++ shared/http-signatures/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @brains/http-signatures
 
+## 0.2.0-alpha.409
+
+## 0.2.0-alpha.408
+
 ## 0.2.0-alpha.407
 
 ## 0.2.0-alpha.406