+++ shared/site-engine/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @brains/site-engine
 
+## 0.2.0-alpha.409
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.409
+  - @brains/image@0.2.0-alpha.409
+  - @brains/site-composition@0.2.0-alpha.409
+  - @brains/ui-library@0.2.0-alpha.409
+  - @brains/utils@0.2.0-alpha.409
+  - @brains/entity-service@0.2.0-alpha.409
+
+## 0.2.0-alpha.408
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/contracts@0.2.0-alpha.408
+  - @brains/image@0.2.0-alpha.408
+  - @brains/site-composition@0.2.0-alpha.408
+  - @brains/ui-library@0.2.0-alpha.408
+  - @brains/utils@0.2.0-alpha.408
+  - @brains/entity-service@0.2.0-alpha.408
+
 ## 0.2.0-alpha.407
 
 ### Patch Changes