+++ sites/professional/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @brains/site-professional
 
+## 0.2.0-alpha.409
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/blog@0.2.0-alpha.409
+  - @brains/decks@0.2.0-alpha.409
+  - @brains/site-info@0.2.0-alpha.409
+  - @brains/topics@0.2.0-alpha.409
+  - @brains/profile@0.2.0-alpha.409
+  - @brains/content-formatters@0.2.0-alpha.409
+  - @brains/contracts@0.2.0-alpha.409
+  - @brains/site-composition@0.2.0-alpha.409
+  - @brains/site-engine@0.2.0-alpha.409
+  - @brains/ui-library@0.2.0-alpha.409
+  - @brains/utils@0.2.0-alpha.409
+  - @brains/plugins@0.2.0-alpha.409
+  - @brains/templates@0.2.0-alpha.409
+
+## 0.2.0-alpha.408
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @brains/blog@0.2.0-alpha.408
+  - @brains/decks@0.2.0-alpha.408
+  - @brains/site-info@0.2.0-alpha.408
+  - @brains/topics@0.2.0-alpha.408
+  - @brains/profile@0.2.0-alpha.408
+  - @brains/content-formatters@0.2.0-alpha.408
+  - @brains/contracts@0.2.0-alpha.408
+  - @brains/site-composition@0.2.0-alpha.408
+  - @brains/site-engine@0.2.0-alpha.408
+  - @brains/ui-library@0.2.0-alpha.408
+  - @brains/utils@0.2.0-alpha.408
+  - @brains/plugins@0.2.0-alpha.408
+  - @brains/templates@0.2.0-alpha.408
+
 ## 0.2.0-alpha.407
 
 ### Patch Changes